primer avance pestana amigos, interfaz amigos

+++ src/controllers/authenticationController.js
@@ -71,16 +71,23 @@ controllerAuthentication.logout = (req, res) => {
 controllerAuthentication.addimage = async (req, res) => {
     try {
         const { id } = req.params
-    
-        if(req.user.public_id!==null){
-            await cloudinary.v2.uploader.destroy(req.user.public_id)
+        if(req.file==undefined){
+            console.log(req.file);
+            req.flash('message', 'Primero seleccionar imagen')
+            req.redirect('/profile')
         }
-        
-        const result = await cloudinary.v2.uploader.upload(req.file.path)
-        await pool.query('update user set imgUser=? , public_id=? where idUser=?', [result.secure_url,result.public_id,id])
-        await fs.unlink(req.file.path)
-        req.flash('success', 'Foto de perfil colocada correctamente')
-        res.redirect('/profile')
+        else{
+
+            if(req.user.public_id!==null){
+                await cloudinary.v2.uploader.destroy(req.user.public_id)
+            }
+            
+            const result = await cloudinary.v2.uploader.upload(req.file.path)
+            await pool.query('update user set imgUser=? , public_id=? where idUser=?', [result.secure_url,result.public_id,id])
+            await fs.unlink(req.file.path)
+            req.flash('success', 'Foto de perfil colocada correctamente')
+            res.redirect('/profile')
+        }
     } catch (e) {
         console.log(e);
     }
